Finally Finished
Everything is done!! Took a month or so to finish the entire game, but with the fifa world cup coming tomorrow, I hope it gets a lot of downloads. Might be even more than I would expect. I will post the game tomorrow and post on the social medias, tagging as many things as I can.